qr-sys/requirements.txt
aiosmtplib==3.0.1
aiosqlite==0.19.0
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async==0.6.2
async-timeout==4.0.3
asyncpg==0.29.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.5.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.0
flower==2.0.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
humanize==4.9.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
Mako==1.3.0
MarkupSafe==2.1.4
orjson==3.9.12
packaging==23.2
pathlib==1.0.1
pillow==10.2.0
pluggy==1.4.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.8.0
pypng==0.20220715.0
PyQRCode==1.2.1
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-order==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
tomli==2.0.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.9.0
uvicorn==0.26.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
